third_party/WebKit/LayoutTests/fast/scroll-behavior/subframe-scrollLeft-expected.txt

Test that setting scrollLeft on a subframe works with both scroll behaviors


This is a testharness.js-based test.
PASS Scroll x:1, y:0, smooth:false 
PASS Scroll x:2, y:0, smooth:false 
PASS Scroll x:3, y:0, smooth:false 
PASS Scroll x:4, y:0, smooth:false 
FAIL Scroll x:10, y:0, smooth:true assert_equals: expected "4, 0" but got "10, 0"
FAIL Scroll x:20, y:0, smooth:true assert_equals: expected "10, 0" but got "20, 0"
FAIL Scroll x:30, y:0, smooth:true assert_equals: expected "20, 0" but got "30, 0"
FAIL Scroll x:40, y:0, smooth:true assert_equals: expected "30, 0" but got "40, 0"
FAIL Scroll x:4000, y:0, smooth:true assert_equals: expected "40, 0" but got "4000, 0"
FAIL Scroll x:15, y:0, smooth:true assert_equals: expected "4000, 0" but got "15, 0"
FAIL Scroll x:4100, y:0, smooth:true assert_equals: expected "15, 0" but got "4100, 0"
FAIL Scroll x:10, y:0, smooth:true assert_equals: expected "4100, 0" but got "10, 0"
Harness: the test ran to completion.
